% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{download_floodpulse}
\alias{download_floodpulse}
\title{Download Sentinel-1 COG from AWS S3}
\usage{
download_floodpulse(s3_path, output_dir = "downloads")
}
\arguments{
\item{s3_path}{S3 path from fetch_floodpulse}

\item{output_dir}{Local directory}
}
\value{
Local file path
}
\description{
Download Sentinel-1 COG from AWS S3
}
