% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{fetch_floodpulse}
\alias{fetch_floodpulse}
\title{Fetch Sentinel-1 GRD Data from AWS S3}
\usage{
fetch_floodpulse(aoi, start_date, end_date)
}
\arguments{
\item{aoi}{Path to shapefile (.shp) defining area of interest}

\item{start_date}{Start date (YYYYMMDD)}

\item{end_date}{End date (YYYYMMDD)}
}
\value{
List of S3 paths to COG files
}
\description{
Fetch Sentinel-1 GRD Data from AWS S3
}
